Move SeasonStats helpers to module
